+++ packages/imperative.ts
@@ -38,6 +38,117 @@ const config: IImperativeConfig = {
             logFile: Constants.LOG_LOCATION
         }
     },
+    baseProfile: {
+        type: "base",
+        schema: {
+            type: "object",
+            title: "Base Profile",
+            description: "Base profile that stores values shared by multiple service profiles",
+            properties: {
+                host: {
+                    type: "string",
+                    optionDefinition: Constants.BASE_OPTION_HOST
+                },
+                port: {
+                    type: "number",
+                    optionDefinition: Constants.BASE_OPTION_PORT
+                },
+                user: {
+                    type: "string",
+                    secure: true,
+                    optionDefinition: Constants.BASE_OPTION_USER
+                },
+                password: {
+                    type: "string",
+                    secure: true,
+                    optionDefinition: Constants.BASE_OPTION_PASSWORD
+                },
+                rejectUnauthorized: {
+                    type: "boolean",
+                    optionDefinition: Constants.BASE_OPTION_REJECT_UNAUTHORIZED
+                },
+                tokenType: {
+                    type: "string",
+                    optionDefinition: Constants.BASE_OPTION_TOKEN_TYPE
+                },
+                tokenValue: {
+                    type: "string",
+                    secure: true,
+                    optionDefinition: Constants.BASE_OPTION_TOKEN_VALUE
+                }
+            },
+            required: []
+        },
+        createProfileExamples: [
+            {
+                options: "base1 --host example.com --port 443 --user admin --password 123456",
+                description: "Create a profile called 'base1' to connect to host example.com and port 443"
+            },
+            {
+                options: "base2 --host example.com --user admin --password 123456 --reject-unauthorized false",
+                description: "Create a profile called 'base2' to connect to host example.com (default port - 443) " +
+                    "and allow self-signed certificates"
+            },
+            {
+                options: "base3 --host example.com --port 1443",
+                description: "Create a profile called 'base3' to connect to host example.com and port 1443, " +
+                    " not specifying a username or password so they are not stored on disk; these will need to be specified on every command"
+            },
+            {
+                options: "base4 --reject-unauthorized false",
+                description: "Create a zosmf profile called 'base4' to connect to default port 443 and allow self-signed certificates, " +
+                    "not specifying a username, password, or host so they are not stored on disk; these will need to be specified on every command"
+            }
+        ],
+        updateProfileExamples: [
+            {
+                options: "base1 --user newuser --password newp4ss",
+                description: "Update a base profile named 'base1' with a new username and password"
+            }
+        ],
+        authConfig: [
+            {
+                serviceName: "apiml",
+                handler: __dirname + "/auth/src/cli/ApimlAuthHandler",
+                login: {
+                    summary: Constants.APIML_LOGIN_SUMMARY,
+                    description: Constants.APIML_LOGIN_DESCRIPTION,
+                    examples: [
+                        Constants.APIML_LOGIN_EXAMPLE1,
+                        Constants.APIML_LOGIN_EXAMPLE2
+                    ],
+                    options: [
+                        Constants.BASE_OPTION_HOST,
+                        Constants.BASE_OPTION_PORT,
+                        Constants.BASE_OPTION_USER,
+                        Constants.BASE_OPTION_PASSWORD,
+                        Constants.BASE_OPTION_REJECT_UNAUTHORIZED
+                    ]
+                },
+                logout: {
+                    summary: Constants.APIML_LOGOUT_SUMMARY,
+                    description: Constants.APIML_LOGOUT_DESCRIPTION,
+                    examples: [
+                        Constants.APIML_LOGOUT_EXAMPLE1,
+                        Constants.APIML_LOGOUT_EXAMPLE2
+                    ],
+                    options: [
+                        Constants.BASE_OPTION_HOST,
+                        Constants.BASE_OPTION_PORT,
+                        Constants.BASE_OPTION_TOKEN_TYPE,
+                        Constants.BASE_OPTION_TOKEN_VALUE,
+                        Constants.BASE_OPTION_REJECT_UNAUTHORIZED
+                    ]
+                }
+            }
+        ]
+    },
+    authGroupConfig: {
+        authGroup: {
+            summary: Constants.AUTH_GROUP_SUMMARY,
+            description: Constants.AUTH_GROUP_DESCRIPTION
+        }
+    },
     profiles: [
         {
             type: "zosmf",
@@ -71,6 +182,16 @@ const config: IImperativeConfig = {
                     basePath: {
                         type: "string",
                         optionDefinition: ZosmfSession.ZOSMF_OPTION_BASE_PATH
+                    },
+                    encoding: {
+                        type: "number",
+                        optionDefinition: {
+                            name: "encoding",
+                            aliases: ["ec"],
+                            description: "The encoding for download and upload of z/OS data set and USS files." +
+                                " The default encoding if not specified is 1047.",
+                            type: "number"
+                        }
                     }
                 },
                 required: []
@@ -93,7 +214,7 @@ const config: IImperativeConfig = {
                 {
                     options: "zos126 --reject-unauthorized false",
                     description: "Create a zosmf profile called 'zos126' to connect to z/OSMF on the default port 443 and allow self-signed certificates, " +
-                    "not specifying a username, password, or host so they are not stored on disk; these will need to be specified on every command"
+                        "not specifying a username, password, or host so they are not stored on disk; these will need to be specified on every command"
                 },
                 {
                     options: "zosAPIML --host zosAPIML --port 2020 --user ibmuser --password myp4ss --reject-unauthorized false --base-path basePath",
@@ -220,13 +341,13 @@ const config: IImperativeConfig = {
                 {
                     options: "ssh333 --host sshhost --user ibmuser --privateKey /path/to/privatekey --keyPassphrase privateKeyPassphrase",
                     description: "Create a ssh profile called 'ssh333' to connect to z/OS SSH server at host 'zos123' " +
-                                 "using a privatekey '/path/to/privatekey' and its decryption passphrase 'privateKeyPassphrase' " +
-                                 "for privatekey authentication"
+                        "using a privatekey '/path/to/privatekey' and its decryption passphrase 'privateKeyPassphrase' " +
+                        "for privatekey authentication"
                 },
                 {
                     options: "ssh444 --privateKey /path/to/privatekey",
                     description: "Create a ssh profile called 'ssh444' to connect to z/OS SSH server on default port 22, without specifying " +
-                    "username, host, or password, preventing those values from being stored on disk"
+                        "username, host, or password, preventing those values from being stored on disk"
                 }
             ]
         }